Refactor `animate` to export a `tween` function
